// src/app/modules/timetracker/enums/category.enum.ts
export enum Category {
  ABSENCE = 'Absence',
  ADMINISTRATIVE = 'Administrative',
  DEVELOPMENT = 'Development',
  DOCUMENTATION = 'Documentation',
  IDLE_TIME = 'Idle time',
  INTERNAL_PROCESS = 'Internal Process',
  MEETINGS_CLIENT = 'Meetings (Client)',
  MEETINGS_INTERNAL = 'Meetings (Internal)',
  OTHER = 'Other',
  TESTING = 'Testing',
  TRAINING_TRAINEE = 'Training (Trainee)',
  TRAINING_TRAINER = 'Training (Trainer)',
}